refactor: slice C combat test support splits (#4196)
Split combat_resolver_test_support builders and remaining >220 phys
scenario tables along existing list seams; every lib file now ≤220 lines.
All 207 colonizethis_combat scenario tests pass.
